#76 customFont 추출 메서드 CustomFont 정의부에 포함
